Move feature tests to a Features subnamespace
