+++ tests/Feature/AssetTest.php
@@ -1,12 +1,13 @@
 <?php
 
-namespace pxgamer\Arionum\Tests;
+namespace pxgamer\Arionum\Tests\Feature;
 
 use GuzzleHttp\Client;
 use GuzzleHttp\HandlerStack;
 use pxgamer\Arionum\Arionum;
 use GuzzleHttp\Psr7\Response;
 use GuzzleHttp\Handler\MockHandler;
+use pxgamer\Arionum\Tests\TestCase;
 use pxgamer\Arionum\Exceptions\GenericApiException;
 
 final class AssetTest extends TestCase
