+++ tests/Feature/EllipticCurveTest.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace pxgamer\Arionum\Tests;
+namespace pxgamer\Arionum\Tests\Feature;
 
+use pxgamer\Arionum\Tests\TestCase;
 use pxgamer\Arionum\Models\Transaction;
 use pxgamer\Arionum\Helpers\EllipticCurve;
 use pxgamer\Arionum\Exceptions\SignatureException;
