+++ tests/Feature/LocalAccountTest.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace pxgamer\Arionum\Tests;
+namespace pxgamer\Arionum\Tests\Feature;
 
 use pxgamer\Arionum\Models\Account;
+use pxgamer\Arionum\Tests\TestCase;
 
 final class LocalAccountTest extends TestCase
 {
