+++ tests/Feature/TransactionTest.php
@@ -2,9 +2,10 @@
 
 declare(strict_types=1);
 
-namespace pxgamer\Arionum\Tests;
+namespace pxgamer\Arionum\Tests\Feature;
 
 use pxgamer\Arionum\Models\Asset;
+use pxgamer\Arionum\Tests\TestCase;
 use pxgamer\Arionum\Models\Transaction;
 use pxgamer\Arionum\Transaction\Version;
 use pxgamer\Arionum\Exceptions\GenericApiException;
